requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-environ==0.4.5
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
mysqlclient==1.4.6
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
uWSGI==2.0.18
